Refactor config variable setup.
Installed python-decouple to help source config values from environment
variables or from an .env file (not being used, currently, but its
another option). The library also has an added value of requiring the
needed environment variables to be set and will throw a descriptive
error for the user if they are not set.

Organized the setting of all all environment variables into the
config.py module.

Refactored parts of the codebase to use these new values created in the
config.py module.