Major refactor to accomodate python's setup.py
This refactor does a few things to bring the project into a more
python focused structure.

The setup.py file that existed was the main script.  Unfortunately,
this is a special file with python's setuptools packager and renders its
use to be confusing.

This patch moves the main logic into a distributable `new_albums` package.
This allows for a convenient `pip install .` for local installation, allowing
the script to be run with the python script runner: ie
`python -m new_albums`.   Additionally, a __main__.py file was
 included to take over the role of the primary script.

Environment variables can also be set in a `.env` file which is ignored from
the version control system.

The data.fiat.py file was awkward to package up.  so I changed the way
we specify custom genres to accept/reject.  By default, the example
file is used.  You can create a new .py file (currently it must be in the
 `new_albums` package, but that should be changed to be more universal.
 You can set the filename using a `FIAT_FILE` envvar.

README also adjusted to match changes.