Deprecate fiat file.
Ensure that the code doesn't reference fiat files anymore. The technical
name for the path where accept.txt and reject.txt reside is now
`fiat_path`. Currently, some of the logic is too tightly coupled
together so that needs to be changed in the future. For example, the
code in `accept_reject` would be better off returning a `list[path]`
instead of a `bool`. However, that's more work than I want to do at the
moment.